+++ src/v/cluster/metadata_dissemination_service.h
@@ -19,11 +19,11 @@
 #include "features/fwd.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "net/unresolved_address.h"
 #include "raft/fwd.h"
 #include "rpc/fwd.h"
 #include "utils/mutex.h"
 #include "utils/retry.h"
+#include "utils/unresolved_address.h"
 
 #include <seastar/core/abort_source.hh>
 #include <seastar/core/sharded.hh>
