+++ src/v/cluster/tests/controller_state_test.cc
@@ -18,9 +18,9 @@
 #include "config/configuration.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
-#include "net/unresolved_address.h"
 #include "test_utils/async.h"
 #include "test_utils/fixture.h"
+#include "utils/unresolved_address.h"
 
 #include <seastar/core/future-util.hh>
 #include <seastar/core/sstring.hh>
