+++ src/v/config/tests/seed_server_property_test.cc
@@ -9,7 +9,7 @@
 
 #include "config/node_config.h"
 #include "config/seed_server.h"
-#include "net/unresolved_address.h"
+#include "utils/unresolved_address.h"
 
 #include <seastar/testing/thread_test_case.hh>
