+++ src/v/kafka/client/test/consumer_group.cc
@@ -36,9 +36,9 @@
 #include "kafka/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "net/unresolved_address.h"
 #include "redpanda/tests/fixture.h"
 #include "ssx/future-util.h"
+#include "utils/unresolved_address.h"
 
 #include <seastar/core/loop.hh>
 #include <seastar/core/sleep.hh>
