+++ src/v/kafka/client/test/fetch.cc
@@ -21,8 +21,8 @@
 #include "kafka/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "net/unresolved_address.h"
 #include "redpanda/tests/fixture.h"
+#include "utils/unresolved_address.h"
 
 #include <boost/test/tools/old/interface.hpp>
