+++ src/v/net/connection_rate.h
@@ -15,10 +15,10 @@
 #include "base/vassert.h"
 #include "config/validators.h"
 #include "net/connection_rate_counter.h"
-#include "net/inet_address_wrapper.h"
 #include "net/server_probe.h"
 #include "seastar/core/coroutine.hh"
 #include "ssx/future-util.h"
+#include "utils/inet_address_wrapper.h"
 
 #include <seastar/core/abort_source.hh>
 #include <seastar/core/gate.hh>
