+++ src/v/net/dns.h
@@ -9,7 +9,7 @@
  * by the Apache License, Version 2.0
  */
 #pragma once
-#include "net/unresolved_address.h"
+#include "utils/unresolved_address.h"
 
 namespace net {
