+++ src/v/pandaproxy/rest/test/consumer_group.cc
@@ -10,12 +10,12 @@
 #include "http/client.h"
 #include "kafka/protocol/join_group.h"
 #include "kafka/types.h"
-#include "net/unresolved_address.h"
 #include "pandaproxy/json/requests/create_consumer.h"
 #include "pandaproxy/json/rjson_util.h"
 #include "pandaproxy/rest/configuration.h"
 #include "pandaproxy/test/pandaproxy_fixture.h"
 #include "pandaproxy/test/utils.h"
+#include "utils/unresolved_address.h"
 
 #include <boost/beast/http/field.hpp>
 #include <boost/beast/http/status.hpp>
